+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New prop `layout` to `filter-navigator.v3`.
+- Allow `drawer` on `search-result-layout.mobile`.
 
 ## [3.39.10] - 2019-12-23
 ### Changed